removed files used for building the model
