--- darts/models/forecasting/tft_pytorch_2.py
@@ -1,839 +1,0 @@
-"""
-The temporal fusion transformer is a powerful predictive model for forecasting timeseries
-"""
-from copy import copy
-from typing import Dict, List, Tuple, Union
-
-from matplotlib import pyplot as plt
-import numpy as np
-import torch
-from torch import nn
-
-from pytorch_forecasting.data import TimeSeriesDataSet
-# from pytorch_forecasting.metrics import MAE, MAPE, MASE, RMSE, SMAPE, MultiHorizonMetric, MultiLoss, QuantileLoss
-from pytorch_forecasting.metrics import MAE, MAPE, MASE, RMSE, SMAPE, MultiHorizonMetric, MultiLoss
-from pytorch_forecasting.models.base_model import BaseModelWithCovariates
-# from pytorch_forecasting.models import LSTM, MultiEmbedding
-from darts.models.forecasting.tft_submodels import (
-    AddNorm,
-    GateAddNorm,
-    GatedLinearUnit,
-    GatedResidualNetwork,
-    InterpretableMultiHeadAttention,
-    VariableSelectionNetwork,
-    LSTM,
-    MultiEmbedding,
-    QuantileLoss
-)
-from pytorch_forecasting.utils import autocorrelation, create_mask, detach, integer_histogram, padded_stack, to_list
-
-
-class TemporalFusionTransformer(BaseModelWithCovariates):
-    def __init__(self,
-                 hidden_size: int = 16,
-                 lstm_layers: int = 1,
-                 dropout: float = 0.1,
-                 output_size: Union[int, List[int]] = 7,
-                 loss: MultiHorizonMetric = None,
-                 attention_head_size: int = 4,
-                 max_encoder_length: int = 10,
-                 static_categoricals: List[str] = [],
-                 static_reals: List[str] = [],
-                 time_varying_categoricals_encoder: List[str] = [],
-                 time_varying_categoricals_decoder: List[str] = [],
-                 categorical_groups: Dict[str, List[str]] = {},
-                 time_varying_reals_encoder: List[str] = [],
-                 time_varying_reals_decoder: List[str] = [],
-                 x_reals: List[str] = [],
-                 x_categoricals: List[str] = [],
-                 hidden_continuous_size: int = 8,
-                 hidden_continuous_sizes: Dict[str, int] = {},
-                 embedding_sizes: Dict[str, Tuple[int, int]] = {},
-                 embedding_paddings: List[str] = [],
-                 embedding_labels: Dict[str, np.ndarray] = {},
-                 learning_rate: float = 1e-3,
-                 log_interval: Union[int, float] = -1,
-                 log_val_interval: Union[int, float] = None,
-                 log_gradient_flow: bool = False,
-                 reduce_on_plateau_patience: int = 1000,
-                 monotone_constaints: Dict[str, int] = {},
-                 share_single_variable_networks: bool = False,
-                 logging_metrics: nn.ModuleList = None,
-                 **kwargs):
-        """
-        Temporal Fusion Transformer for forecasting timeseries - use its :py:meth:`~from_dataset` method if possible.
-
-        Implementation of the article
-        `Temporal Fusion Transformers for Interpretable Multi-horizon Time Series
-        Forecasting <https://arxiv.org/pdf/1912.09363.pdf>`_. The network outperforms DeepAR by Amazon by 36-69%
-        in benchmarks.
-
-        Enhancements compared to the original implementation (apart from capabilities added through base model
-        such as monotone constraints):
-
-        * static variables can be continuous
-        * multiple categorical variables can be summarized with an EmbeddingBag
-        * variable encoder and decoder length by sample
-        * categorical embeddings are not transformed by variable selection network (because it is a redundant operation)
-        * variable dimension in variable selection network are scaled up via linear interpolation to reduce
-          number of parameters
-        * non-linear variable processing in variable selection network can be shared among decoder and encoder
-          (not shared by default)
-
-        Tune its hyperparameters with
-        :py:func:`~pytorch_forecasting.models.temporal_fusion_transformer.tuning.optimize_hyperparameters`.
-
-        Args:
-
-            hidden_size: hidden size of network which is its main hyperparameter and can range from 8 to 512
-            lstm_layers: number of LSTM layers (2 is mostly optimal)
-            dropout: dropout rate
-            output_size: number of outputs (e.g. number of quantiles for QuantileLoss and one target or list
-                of output sizes).
-            loss: loss function taking prediction and targets
-            attention_head_size: number of attention heads (4 is a good default)
-            max_encoder_length: length to encode (can be far longer than the decoder length but does not have to be)
-            static_categoricals: names of static categorical variables
-            static_reals: names of static continuous variables
-            time_varying_categoricals_encoder: names of categorical variables for encoder
-            time_varying_categoricals_decoder: names of categorical variables for decoder
-            time_varying_reals_encoder: names of continuous variables for encoder
-            time_varying_reals_decoder: names of continuous variables for decoder
-            categorical_groups: dictionary where values
-                are list of categorical variables that are forming together a new categorical
-                variable which is the key in the dictionary
-            x_reals: order of continuous variables in tensor passed to forward function
-            x_categoricals: order of categorical variables in tensor passed to forward function
-            hidden_continuous_size: default for hidden size for processing continous variables (similar to categorical
-                embedding size)
-            hidden_continuous_sizes: dictionary mapping continuous input indices to sizes for variable selection
-                (fallback to hidden_continuous_size if index is not in dictionary)
-            embedding_sizes: dictionary mapping (string) indices to tuple of number of categorical classes and
-                embedding size
-            embedding_paddings: list of indices for embeddings which transform the zero's embedding to a zero vector
-            embedding_labels: dictionary mapping (string) indices to list of categorical labels
-            learning_rate: learning rate
-            log_interval: log predictions every x batches, do not log if 0 or less, log interpretation if > 0. If < 1.0
-                , will log multiple entries per batch. Defaults to -1.
-            log_val_interval: frequency with which to log validation set metrics, defaults to log_interval
-            log_gradient_flow: if to log gradient flow, this takes time and should be only done to diagnose training
-                failures
-            reduce_on_plateau_patience (int): patience after which learning rate is reduced by a factor of 10
-            monotone_constaints (Dict[str, int]): dictionary of monotonicity constraints for continuous decoder
-                variables mapping
-                position (e.g. ``"0"`` for first position) to constraint (``-1`` for negative and ``+1`` for positive,
-                larger numbers add more weight to the constraint vs. the loss but are usually not necessary).
-                This constraint significantly slows down training. Defaults to {}.
-            share_single_variable_networks (bool): if to share the single variable networks between the encoder and
-                decoder. Defaults to False.
-            logging_metrics (nn.ModuleList[LightningMetric]): list of metrics that are logged during training.
-                Defaults to nn.ModuleList([SMAPE(), MAE(), RMSE(), MAPE()]).
-            **kwargs: additional arguments to :py:class:`~BaseModel`.
-        """
-        if logging_metrics is None:
-            logging_metrics = nn.ModuleList([SMAPE(), MAE(), RMSE(), MAPE()])
-        if loss is None:
-            loss = QuantileLoss()
-        # # store loss function separately as it is a module
-        # assert isinstance(loss, LightningMetric), "Loss has to be a PyTorch Lightning `Metric`"
-        self.save_hyperparameters()
-        super().__init__(loss=loss, logging_metrics=logging_metrics, **kwargs)
-
-        self.hidden_size = hidden_size
-        self.lstm_layers = lstm_layers
-        self.dropout = dropout
-        self.output_size = output_size
-        self.loss = loss
-        self.attention_head_size = attention_head_size
-        self.max_encoder_length = max_encoder_length
-        self.static_categoricals = static_categoricals
-        self.static_reals = static_reals
-        self.time_varying_categoricals_encoder = time_varying_categoricals_encoder
-        self.time_varying_categoricals_decoder = time_varying_categoricals_decoder
-        self.categorical_groups = categorical_groups
-        self.time_varying_reals_encoder = time_varying_reals_encoder
-        self.time_varying_reals_decoder = time_varying_reals_decoder
-        self.x_reals = x_reals
-        self.x_categoricals = x_categoricals
-        self.hidden_continuous_size = hidden_continuous_size
-        self.hidden_continuous_sizes = hidden_continuous_sizes
-        self.embedding_sizes = embedding_sizes
-        self.embedding_paddings = embedding_paddings
-        self.embedding_labels = embedding_labels
-        self.learning_rate = learning_rate
-        self.log_interval = log_interval
-        self.log_val_interval = log_val_interval
-        self.log_gradient_flow = log_gradient_flow
-        self.reduce_on_plateau_patience = reduce_on_plateau_patience
-        self.monotone_constaints = monotone_constaints
-        self.share_single_variable_networks = share_single_variable_networks
-        self.logging_metrics = logging_metrics
-        self.kwargs = kwargs
-
-        # processing inputs
-        # embeddings
-        self.input_embeddings = MultiEmbedding(
-            embedding_sizes=self.embedding_sizes,
-            categorical_groups=self.categorical_groups,
-            embedding_paddings=self.embedding_paddings,
-            x_categoricals=self.x_categoricals,
-            max_embedding_size=self.hidden_size,
-        )
-
-        # update embedding sizes
-        self.embedding_sizes = self.input_embeddings.embedding_sizes
-
-        # continuous variable processing
-        self.prescalers = nn.ModuleDict(
-            {
-                name: nn.Linear(1, self.hidden_continuous_sizes.get(name, self.hidden_continuous_size))
-                for name in self.reals
-            }
-        )
-
-        # variable selection
-        # variable selection for static variables
-        static_input_sizes = {name: self.input_embeddings.embedding_sizes[name][1] for name in self.static_categoricals}
-        static_input_sizes.update(
-            {
-                name: self.hidden_continuous_sizes.get(name, self.hidden_continuous_size)
-                for name in self.static_reals
-            }
-        )
-        self.static_variable_selection = VariableSelectionNetwork(
-            input_sizes=static_input_sizes,
-            hidden_size=self.hidden_size,
-            input_embedding_flags={name: True for name in self.static_categoricals},
-            dropout=self.dropout,
-            prescalers=self.prescalers,
-        )
-
-        # variable selection for encoder and decoder
-        encoder_input_sizes = {
-            name: self.input_embeddings.embedding_sizes[name][1] for name in self.time_varying_categoricals_encoder
-        }
-        encoder_input_sizes.update(
-            {
-                name: self.hidden_continuous_sizes.get(name, self.hidden_continuous_size)
-                for name in self.time_varying_reals_encoder
-            }
-        )
-
-        decoder_input_sizes = {
-            name: self.input_embeddings.embedding_sizes[name][1] for name in self.time_varying_categoricals_decoder
-        }
-        decoder_input_sizes.update(
-            {
-                name: self.hidden_continuous_sizes.get(name, self.hidden_continuous_size)
-                for name in self.time_varying_reals_decoder
-            }
-        )
-
-        # create single variable grns that are shared across decoder and encoder
-        if self.share_single_variable_networks:
-            self.shared_single_variable_grns = nn.ModuleDict()
-            for name, input_size in encoder_input_sizes.items():
-                self.shared_single_variable_grns[name] = GatedResidualNetwork(
-                    input_size,
-                    min(input_size, self.hidden_size),
-                    self.hidden_size,
-                    self.dropout,
-                )
-            for name, input_size in decoder_input_sizes.items():
-                if name not in self.shared_single_variable_grns:
-                    self.shared_single_variable_grns[name] = GatedResidualNetwork(
-                        input_size,
-                        min(input_size, self.hidden_size),
-                        self.hidden_size,
-                        self.dropout,
-                    )
-
-        self.encoder_variable_selection = VariableSelectionNetwork(
-            input_sizes=encoder_input_sizes,
-            hidden_size=self.hidden_size,
-            input_embedding_flags={name: True for name in self.time_varying_categoricals_encoder},
-            dropout=self.dropout,
-            context_size=self.hidden_size,
-            prescalers=self.prescalers,
-            single_variable_grns={}
-            if not self.share_single_variable_networks
-            else self.shared_single_variable_grns,
-        )
-
-        self.decoder_variable_selection = VariableSelectionNetwork(
-            input_sizes=decoder_input_sizes,
-            hidden_size=self.hidden_size,
-            input_embedding_flags={name: True for name in self.time_varying_categoricals_decoder},
-            dropout=self.dropout,
-            context_size=self.hidden_size,
-            prescalers=self.prescalers,
-            single_variable_grns={}
-            if not self.share_single_variable_networks
-            else self.shared_single_variable_grns,
-        )
-
-        # static encoders
-        # for variable selection
-        self.static_context_variable_selection = GatedResidualNetwork(
-            input_size=self.hidden_size,
-            hidden_size=self.hidden_size,
-            output_size=self.hidden_size,
-            dropout=self.dropout,
-        )
-
-        # for hidden state of the lstm
-        self.static_context_initial_hidden_lstm = GatedResidualNetwork(
-            input_size=self.hidden_size,
-            hidden_size=self.hidden_size,
-            output_size=self.hidden_size,
-            dropout=self.dropout,
-        )
-
-        # for cell state of the lstm
-        self.static_context_initial_cell_lstm = GatedResidualNetwork(
-            input_size=self.hidden_size,
-            hidden_size=self.hidden_size,
-            output_size=self.hidden_size,
-            dropout=self.dropout,
-        )
-
-        # for post lstm static enrichment
-        self.static_context_enrichment = GatedResidualNetwork(
-            self.hidden_size, self.hidden_size, self.hidden_size, self.dropout
-        )
-
-        # lstm encoder (history) and decoder (future) for local processing
-        self.lstm_encoder = LSTM(
-            input_size=self.hidden_size,
-            hidden_size=self.hidden_size,
-            num_layers=self.lstm_layers,
-            dropout=self.dropout if self.lstm_layers > 1 else 0,
-            batch_first=True,
-        )
-
-        self.lstm_decoder = LSTM(
-            input_size=self.hidden_size,
-            hidden_size=self.hidden_size,
-            num_layers=self.lstm_layers,
-            dropout=self.dropout if self.lstm_layers > 1 else 0,
-            batch_first=True,
-        )
-
-        # skip connection for lstm
-        self.post_lstm_gate_encoder = GatedLinearUnit(self.hidden_size, dropout=self.dropout)
-        self.post_lstm_gate_decoder = self.post_lstm_gate_encoder
-        self.post_lstm_add_norm_encoder = AddNorm(self.hidden_size)
-        self.post_lstm_add_norm_decoder = self.post_lstm_add_norm_encoder
-
-        # static enrichment and processing past LSTM
-        self.static_enrichment = GatedResidualNetwork(
-            input_size=self.hidden_size,
-            hidden_size=self.hidden_size,
-            output_size=self.hidden_size,
-            dropout=self.dropout,
-            context_size=self.hidden_size,
-        )
-
-        # attention for long-range processing
-        self.multihead_attn = InterpretableMultiHeadAttention(
-            d_model=self.hidden_size, n_head=self.attention_head_size, dropout=self.dropout
-        )
-        self.post_attn_gate_norm = GateAddNorm(self.hidden_size, dropout=self.dropout)
-        self.pos_wise_ff = GatedResidualNetwork(
-            self.hidden_size, self.hidden_size, self.hidden_size, dropout=self.dropout
-        )
-
-        # output processing -> no dropout at this late stage
-        self.pre_output_gate_norm = GateAddNorm(self.hidden_size, dropout=None)
-
-        if self.n_targets > 1:  # if to run with multiple targets
-            self.output_layer = nn.ModuleList(
-                [nn.Linear(self.hidden_size, output_size) for output_size in self.output_size]
-            )
-        else:
-            self.output_layer = nn.Linear(self.hidden_size, self.output_size)
-
-    @classmethod
-    def from_dataset(
-        cls,
-        dataset: TimeSeriesDataSet,
-        allowed_encoder_known_variable_names: List[str] = None,
-        **kwargs,
-    ):
-        """
-        Create model from dataset.
-
-        Args:
-            dataset: timeseries dataset
-            allowed_encoder_known_variable_names: List of known variables that are allowed in encoder, defaults to all
-            **kwargs: additional arguments such as hyperparameters for model (see ``__init__()``)
-
-        Returns:
-            TemporalFusionTransformer
-        """
-        # add maximum encoder length
-        # update defaults
-        new_kwargs = copy(kwargs)
-        new_kwargs["max_encoder_length"] = dataset.max_encoder_length
-        new_kwargs.update(cls.deduce_default_output_parameters(dataset, kwargs, QuantileLoss()))
-
-        # create class and return
-        return super().from_dataset(
-            dataset, allowed_encoder_known_variable_names=allowed_encoder_known_variable_names, **new_kwargs
-        )
-
-    def expand_static_context(self, context, timesteps):
-        """
-        add time dimension to static context
-        """
-        return context[:, None].expand(-1, timesteps, -1)
-
-    def get_attention_mask(self, encoder_lengths: torch.LongTensor, decoder_length: int):
-        """
-        Returns causal mask to apply for self-attention layer.
-
-        Args:
-            self_attn_inputs: Inputs to self attention layer to determine mask shape
-        """
-        # indices to which is attended
-        attend_step = torch.arange(decoder_length, device=self.device)
-        # indices for which is predicted
-        predict_step = torch.arange(0, decoder_length, device=self.device)[:, None]
-        # do not attend to steps to self or after prediction
-        # todo: there is potential value in attending to future forecasts if they are made with knowledge currently
-        #   available
-        #   one possibility is here to use a second attention layer for future attention (assuming different effects
-        #   matter in the future than the past)
-        #   or alternatively using the same layer but allowing forward attention - i.e. only masking out non-available
-        #   data and self
-        decoder_mask = attend_step >= predict_step
-        # do not attend to steps where data is padded
-        encoder_mask = create_mask(encoder_lengths.max(), encoder_lengths)
-        # combine masks along attended time - first encoder and then decoder
-        mask = torch.cat(
-            (
-                encoder_mask.unsqueeze(1).expand(-1, decoder_length, -1),
-                decoder_mask.unsqueeze(0).expand(encoder_lengths.size(0), -1, -1),
-            ),
-            dim=2,
-        )
-        return mask
-
-    def forward(self, x: Dict[str, torch.Tensor]) -> Dict[str, torch.Tensor]:
-        """
-        input dimensions: n_samples x time x variables
-        """
-        encoder_lengths = x["encoder_lengths"]
-        decoder_lengths = x["decoder_lengths"]
-        x_cat = torch.cat([x["encoder_cat"], x["decoder_cat"]], dim=1)  # concatenate in time dimension
-        x_cont = torch.cat([x["encoder_cont"], x["decoder_cont"]], dim=1)  # concatenate in time dimension
-        timesteps = x_cont.size(1)  # encode + decode length
-        max_encoder_length = int(encoder_lengths.max())
-        input_vectors = self.input_embeddings(x_cat)
-        input_vectors.update(
-            {
-                name: x_cont[..., idx].unsqueeze(-1)
-                for idx, name in enumerate(self.x_reals)
-                if name in self.reals
-            }
-        )
-
-        # Embedding and variable selection
-        if len(self.static_variables) > 0:
-            # static embeddings will be constant over entire batch
-            static_embedding = {name: input_vectors[name][:, 0] for name in self.static_variables}
-            static_embedding, static_variable_selection = self.static_variable_selection(static_embedding)
-        else:
-            static_embedding = torch.zeros(
-                (x_cont.size(0), self.hidden_size), dtype=self.dtype, device=self.device
-            )
-            static_variable_selection = torch.zeros((x_cont.size(0), 0), dtype=self.dtype, device=self.device)
-
-        static_context_variable_selection = self.expand_static_context(
-            self.static_context_variable_selection(static_embedding), timesteps
-        )
-
-        embeddings_varying_encoder = {
-            name: input_vectors[name][:, :max_encoder_length] for name in self.encoder_variables
-        }
-        embeddings_varying_encoder, encoder_sparse_weights = self.encoder_variable_selection(
-            embeddings_varying_encoder,
-            static_context_variable_selection[:, :max_encoder_length],
-        )
-
-        embeddings_varying_decoder = {
-            name: input_vectors[name][:, max_encoder_length:] for name in self.decoder_variables  # select decoder
-        }
-        embeddings_varying_decoder, decoder_sparse_weights = self.decoder_variable_selection(
-            embeddings_varying_decoder,
-            static_context_variable_selection[:, max_encoder_length:],
-        )
-
-        # LSTM
-        # calculate initial state
-        input_hidden = self.static_context_initial_hidden_lstm(static_embedding).expand(
-            self.lstm_layers, -1, -1
-        )
-        input_cell = self.static_context_initial_cell_lstm(static_embedding).expand(self.lstm_layers, -1, -1)
-
-        # run local encoder
-        encoder_output, (hidden, cell) = self.lstm_encoder(
-            embeddings_varying_encoder, (input_hidden, input_cell), lengths=encoder_lengths, enforce_sorted=False
-        )
-
-        # run local decoder
-        decoder_output, _ = self.lstm_decoder(
-            embeddings_varying_decoder,
-            (hidden, cell),
-            lengths=decoder_lengths,
-            enforce_sorted=False,
-        )
-
-        # skip connection over lstm
-        lstm_output_encoder = self.post_lstm_gate_encoder(encoder_output)
-        lstm_output_encoder = self.post_lstm_add_norm_encoder(lstm_output_encoder, embeddings_varying_encoder)
-
-        lstm_output_decoder = self.post_lstm_gate_decoder(decoder_output)
-        lstm_output_decoder = self.post_lstm_add_norm_decoder(lstm_output_decoder, embeddings_varying_decoder)
-
-        lstm_output = torch.cat([lstm_output_encoder, lstm_output_decoder], dim=1)
-
-        # static enrichment
-        static_context_enrichment = self.static_context_enrichment(static_embedding)
-        attn_input = self.static_enrichment(
-            lstm_output, self.expand_static_context(static_context_enrichment, timesteps)
-        )
-
-        # Attention
-        attn_output, attn_output_weights = self.multihead_attn(
-            q=attn_input[:, max_encoder_length:],  # query only for predictions
-            k=attn_input,
-            v=attn_input,
-            mask=self.get_attention_mask(
-                encoder_lengths=encoder_lengths, decoder_length=timesteps - max_encoder_length
-            ),
-        )
-
-        # skip connection over attention
-        attn_output = self.post_attn_gate_norm(attn_output, attn_input[:, max_encoder_length:])
-
-        output = self.pos_wise_ff(attn_output)
-
-        # skip connection over temporal fusion decoder (not LSTM decoder despite the LSTM output contains
-        # a skip from the variable selection network)
-        output = self.pre_output_gate_norm(output, lstm_output[:, max_encoder_length:])
-        if self.n_targets > 1:  # if to use multi-target architecture
-            output = [output_layer(output) for output_layer in self.output_layer]
-        else:
-            output = self.output_layer(output)
-
-        return self.to_network_output(
-            prediction=self.transform_output(output, target_scale=x["target_scale"]),
-            attention=attn_output_weights,
-            static_variables=static_variable_selection,
-            encoder_variables=encoder_sparse_weights,
-            decoder_variables=decoder_sparse_weights,
-            decoder_lengths=decoder_lengths,
-            encoder_lengths=encoder_lengths,
-        )
-
-    def on_fit_end(self):
-        if self.log_interval > 0:
-            self.log_embeddings()
-
-    def create_log(self, x, y, out, batch_idx, **kwargs):
-        log = super().create_log(x, y, out, batch_idx, **kwargs)
-        if self.log_interval > 0:
-            log["interpretation"] = self._log_interpretation(out)
-        return log
-
-    def _log_interpretation(self, out):
-        # calculate interpretations etc for latter logging
-        interpretation = self.interpret_output(
-            detach(out),
-            reduction="sum",
-            attention_prediction_horizon=0,  # attention only for first prediction horizon
-        )
-        return interpretation
-
-    def epoch_end(self, outputs):
-        """
-        run at epoch end for training or validation
-        """
-        if self.log_interval > 0:
-            self.log_interpretation(outputs)
-
-    def interpret_output(
-        self,
-        out: Dict[str, torch.Tensor],
-        reduction: str = "none",
-        attention_prediction_horizon: int = 0,
-        attention_as_autocorrelation: bool = False,
-    ) -> Dict[str, torch.Tensor]:
-        """
-        interpret output of model
-
-        Args:
-            out: output as produced by ``forward()``
-            reduction: "none" for no averaging over batches, "sum" for summing attentions, "mean" for
-                normalizing by encode lengths
-            attention_prediction_horizon: which prediction horizon to use for attention
-            attention_as_autocorrelation: if to record attention as autocorrelation - this should be set to true in
-                case of ``reduction != "none"`` and differing prediction times of the samples. Defaults to False
-
-        Returns:
-            interpretations that can be plotted with ``plot_interpretation()``
-        """
-
-        # histogram of decode and encode lengths
-        encoder_length_histogram = integer_histogram(out["encoder_lengths"], min=0, max=self.max_encoder_length)
-        decoder_length_histogram = integer_histogram(
-            out["decoder_lengths"], min=1, max=out["decoder_variables"].size(1)
-        )
-
-        # mask where decoder and encoder where not applied when averaging variable selection weights
-        encoder_variables = out["encoder_variables"].squeeze(-2)
-        encode_mask = create_mask(encoder_variables.size(1), out["encoder_lengths"])
-        encoder_variables = encoder_variables.masked_fill(encode_mask.unsqueeze(-1), 0.0).sum(dim=1)
-        encoder_variables /= (
-            out["encoder_lengths"]
-            .where(out["encoder_lengths"] > 0, torch.ones_like(out["encoder_lengths"]))
-            .unsqueeze(-1)
-        )
-
-        decoder_variables = out["decoder_variables"].squeeze(-2)
-        decode_mask = create_mask(decoder_variables.size(1), out["decoder_lengths"])
-        decoder_variables = decoder_variables.masked_fill(decode_mask.unsqueeze(-1), 0.0).sum(dim=1)
-        decoder_variables /= out["decoder_lengths"].unsqueeze(-1)
-
-        # static variables need no masking
-        static_variables = out["static_variables"].squeeze(1)
-        # attention is batch x time x heads x time_to_attend
-        # average over heads + only keep prediction attention and attention on observed timesteps
-        attention = out["attention"][
-            :, attention_prediction_horizon, :, : out["encoder_lengths"].max() + attention_prediction_horizon
-        ].mean(1)
-
-        if reduction != "none":  # if to average over batches
-            static_variables = static_variables.sum(dim=0)
-            encoder_variables = encoder_variables.sum(dim=0)
-            decoder_variables = decoder_variables.sum(dim=0)
-
-            # reorder attention or averaging
-            for i in range(len(attention)):  # very inefficient but does the trick
-                if 0 < out["encoder_lengths"][i] < attention.size(1) - attention_prediction_horizon - 1:
-                    relevant_attention = attention[
-                        i, : out["encoder_lengths"][i] + attention_prediction_horizon
-                    ].clone()
-                    if attention_as_autocorrelation:
-                        relevant_attention = autocorrelation(relevant_attention)
-                    attention[i, -out["encoder_lengths"][i] - attention_prediction_horizon :] = relevant_attention
-                    attention[i, : attention.size(1) - out["encoder_lengths"][i] - attention_prediction_horizon] = 0.0
-                elif attention_as_autocorrelation:
-                    attention[i] = autocorrelation(attention[i])
-
-            attention = attention.sum(dim=0)
-            if reduction == "mean":
-                attention = attention / encoder_length_histogram[1:].flip(0).cumsum(0).clamp(1)
-                attention = attention / attention.sum(-1).unsqueeze(-1)  # renormalize
-            elif reduction == "sum":
-                pass
-            else:
-                raise ValueError(f"Unknown reduction {reduction}")
-
-            attention = torch.zeros(
-                self.max_encoder_length + attention_prediction_horizon, device=self.device
-            ).scatter(
-                dim=0,
-                index=torch.arange(
-                    self.max_encoder_length + attention_prediction_horizon - attention.size(-1),
-                    self.max_encoder_length + attention_prediction_horizon,
-                    device=self.device,
-                ),
-                src=attention,
-            )
-        else:
-            attention = attention / attention.sum(-1).unsqueeze(-1)  # renormalize
-
-        interpretation = dict(
-            attention=attention,
-            static_variables=static_variables,
-            encoder_variables=encoder_variables,
-            decoder_variables=decoder_variables,
-            encoder_length_histogram=encoder_length_histogram,
-            decoder_length_histogram=decoder_length_histogram,
-        )
-        return interpretation
-
-    def plot_prediction(
-        self,
-        x: Dict[str, torch.Tensor],
-        out: Dict[str, torch.Tensor],
-        idx: int,
-        plot_attention: bool = True,
-        add_loss_to_title: bool = False,
-        show_future_observed: bool = True,
-        ax=None,
-        **kwargs,
-    ) -> plt.Figure:
-        """
-        Plot actuals vs prediction and attention
-
-        Args:
-            x (Dict[str, torch.Tensor]): network input
-            out (Dict[str, torch.Tensor]): network output
-            idx (int): sample index
-            plot_attention: if to plot attention on secondary axis
-            add_loss_to_title: if to add loss to title. Default to False.
-            show_future_observed: if to show actuals for future. Defaults to True.
-            ax: matplotlib axes to plot on
-
-        Returns:
-            plt.Figure: matplotlib figure
-        """
-
-        # plot prediction as normal
-        fig = super().plot_prediction(
-            x,
-            out,
-            idx=idx,
-            add_loss_to_title=add_loss_to_title,
-            show_future_observed=show_future_observed,
-            ax=ax,
-            **kwargs,
-        )
-
-        # add attention on secondary axis
-        if plot_attention:
-            interpretation = self.interpret_output(out)
-            for f in to_list(fig):
-                ax = f.axes[0]
-                ax2 = ax.twinx()
-                ax2.set_ylabel("Attention")
-                encoder_length = x["encoder_lengths"][idx]
-                ax2.plot(
-                    torch.arange(-encoder_length, 0),
-                    interpretation["attention"][idx, :encoder_length].detach().cpu(),
-                    alpha=0.2,
-                    color="k",
-                )
-                f.tight_layout()
-        return fig
-
-    def plot_interpretation(self, interpretation: Dict[str, torch.Tensor]) -> Dict[str, plt.Figure]:
-        """
-        Make figures that interpret model.
-
-        * Attention
-        * Variable selection weights / importances
-
-        Args:
-            interpretation: as obtained from ``interpret_output()``
-
-        Returns:
-            dictionary of matplotlib figures
-        """
-        figs = {}
-
-        # attention
-        fig, ax = plt.subplots()
-        attention = interpretation["attention"].detach().cpu()
-        attention = attention / attention.sum(-1).unsqueeze(-1)
-        ax.plot(
-            np.arange(-self.max_encoder_length, attention.size(0) - self.max_encoder_length), attention
-        )
-        ax.set_xlabel("Time index")
-        ax.set_ylabel("Attention")
-        ax.set_title("Attention")
-        figs["attention"] = fig
-
-        # variable selection
-        def make_selection_plot(title, values, labels):
-            fig, ax = plt.subplots(figsize=(7, len(values) * 0.25 + 2))
-            order = np.argsort(values)
-            values = values / values.sum(-1).unsqueeze(-1)
-            ax.barh(np.arange(len(values)), values[order] * 100, tick_label=np.asarray(labels)[order])
-            ax.set_title(title)
-            ax.set_xlabel("Importance in %")
-            plt.tight_layout()
-            return fig
-
-        figs["static_variables"] = make_selection_plot(
-            "Static variables importance", interpretation["static_variables"].detach().cpu(), self.static_variables
-        )
-        figs["encoder_variables"] = make_selection_plot(
-            "Encoder variables importance", interpretation["encoder_variables"].detach().cpu(), self.encoder_variables
-        )
-        figs["decoder_variables"] = make_selection_plot(
-            "Decoder variables importance", interpretation["decoder_variables"].detach().cpu(), self.decoder_variables
-        )
-
-        return figs
-
-    def log_interpretation(self, outputs):
-        """
-        Log interpretation metrics to tensorboard.
-        """
-        # extract interpretations
-        interpretation = {
-            # use padded_stack because decoder length histogram can be of different length
-            name: padded_stack([x["interpretation"][name].detach() for x in outputs], side="right", value=0).sum(0)
-            for name in outputs[0]["interpretation"].keys()
-        }
-        # normalize attention with length histogram squared to account for: 1. zeros in attention and
-        # 2. higher attention due to less values
-        attention_occurances = interpretation["encoder_length_histogram"][1:].flip(0).cumsum(0).float()
-        attention_occurances = attention_occurances / attention_occurances.max()
-        attention_occurances = torch.cat(
-            [
-                attention_occurances,
-                torch.ones(
-                    interpretation["attention"].size(0) - attention_occurances.size(0),
-                    dtype=attention_occurances.dtype,
-                    device=attention_occurances.device,
-                ),
-            ],
-            dim=0,
-        )
-        interpretation["attention"] = interpretation["attention"] / attention_occurances.pow(2).clamp(1.0)
-        interpretation["attention"] = interpretation["attention"] / interpretation["attention"].sum()
-
-        figs = self.plot_interpretation(interpretation)  # make interpretation figures
-        label = ["val", "train"][self.training]
-        # log to tensorboard
-        for name, fig in figs.items():
-            self.logger.experiment.add_figure(
-                f"{label.capitalize()} {name} importance", fig, global_step=self.global_step
-            )
-
-        # log lengths of encoder/decoder
-        for type in ["encoder", "decoder"]:
-            fig, ax = plt.subplots()
-            lengths = (
-                padded_stack([out["interpretation"][f"{type}_length_histogram"] for out in outputs])
-                .sum(0)
-                .detach()
-                .cpu()
-            )
-            if type == "decoder":
-                start = 1
-            else:
-                start = 0
-            ax.plot(torch.arange(start, start + len(lengths)), lengths)
-            ax.set_xlabel(f"{type.capitalize()} length")
-            ax.set_ylabel("Number of samples")
-            ax.set_title(f"{type.capitalize()} length distribution in {label} epoch")
-
-            self.logger.experiment.add_figure(
-                f"{label.capitalize()} {type} length distribution", fig, global_step=self.global_step
-            )
-
-    def log_embeddings(self):
-        """
-        Log embeddings to tensorboard
-        """
-        for name, emb in self.input_embeddings.items():
-            labels = self.embedding_labels[name]
-            self.logger.experiment.add_embedding(
-                emb.weight.data.detach().cpu(), metadata=labels, tag=name, global_step=self.global_step
-            )
