+++ darts/models/forecasting/tft_submodels.py
@@ -9,43 +9,76 @@ import torch.nn as nn
 import torch.nn.functional as F
 from torch.nn.utils import rnn
 
-# # #
+from darts.logging import get_logger, raise_if_not
 
-"""
-Implementations of flexible GRU and LSTM that can handle sequences of length 0.
-"""
+logger = get_logger(__name__)
 
 HiddenState = Union[Tuple[torch.Tensor, torch.Tensor], torch.Tensor]
-
 
 USE_ADAPTION_GLU = False  # seems to work better without
 USE_ADAPTION_RESAMPLE = False  # seems to work better without
 USE_ADAPTION_VSN = False  # seems to work better without
 USE_ADAPTION_NO_ATTENTION_DROPOUT = False  # seems to work better without
+USE_ADAPTION_GOOGLE = False
 
 
 class QuantileLoss(nn.Module):
     """From: https://medium.com/the-artificial-impostor/quantile-regression-part-2-6fdbc26b2629"""
 
-    def __init__(self,
-                 quantiles: Optional[List[float]] = None):
+    def __init__(self, quantiles: Optional[List[float]] = None):
         """
         Arguments:
             quantiles: list of quantiles
         """
+
         super().__init__()
         self.quantiles = [0.02, 0.1, 0.25, 0.5, 0.75, 0.9, 0.98] if quantiles is None else quantiles
+        self.check_quantiles(self.quantiles)
+        self.first = True
 
     def forward(self, y_pred, y_true):
+        """
+        Parameters
+        ----------
+        y_pred
+            must be of shape (n_samples, n_timesteps, n_target_variables, n_quantiles)
+        y_true
+            must be of shape (n_samples, n_timesteps, n_target_variables)
+        """
+
+        dim_q = 3
+
+        if self.first:  # test if torch model forward produces correct output
+            raise_if_not(len(y_pred.shape) == 4 or len(y_true.shape) == 3 or y_pred.shape[:2] != y_true.shape[:2],
+                         'mismatch between predicted and target shape',
+                         logger)
+            raise_if_not(y_pred.shape[dim_q] == len(self.quantiles),
+                         'mismatch between number of predicted quantiles and target quantiles',
+                         logger)
+            self.first = False
+
         losses = []
         for i, q in enumerate(self.quantiles):
-            errors = y_true - y_pred[:, i]
-            losses.append(torch.max((q - 1) * errors, q * errors).unsqueeze(1))
-        loss = torch.mean(torch.sum(torch.cat(losses, dim=1), dim=1))
-        return loss
+            errors = y_true - y_pred[..., i]
+            losses.append(torch.max((q - 1) * errors, q * errors).unsqueeze(dim_q))
+        losses = torch.cat(losses, dim=dim_q)
+        return losses.sum(dim=dim_q).mean()
+
+    @staticmethod
+    def check_quantiles(quantiles):
+        median_q = 0.5
+        raise_if_not(median_q in quantiles,
+                     'median quantile `q=0.5` must be in `quantiles`',
+                     logger)
+        is_centered = [(median_q - left_q) == -(median_q - right_q)
+                       for left_q, right_q in zip(quantiles, quantiles[::-1])]
+        raise_if_not(all(is_centered),
+                     'quantiles lower than `q=0.5` need to share same difference to `0.5` as quantiles '
+                     'higher than `q=0.5`',
+                     logger)
 
 
-class RNN(ABC, nn.RNNBase):
+class _RNN(ABC, nn.RNNBase):
     """
     Base class flexible RNNs.
 
@@ -68,6 +101,7 @@ class RNN(ABC, nn.RNNBase):
         Returns:
             HiddenState: hidden state with propagated initial hidden state where appropriate
         """
+
         pass
 
     @abstractmethod
@@ -82,6 +116,7 @@ class RNN(ABC, nn.RNNBase):
         Returns:
             HiddenState: default (zero-like) hidden state
         """
+
         pass
 
     @abstractmethod
@@ -98,6 +133,7 @@ class RNN(ABC, nn.RNNBase):
         Returns:
             HiddenState: repeated hidden state
         """
+
         pass
 
     def forward(self,
@@ -123,6 +159,7 @@ class RNN(ABC, nn.RNNBase):
             Tuple[Union[rnn.PackedSequence, torch.Tensor], HiddenState]: output and hidden state.
                 Output is packed sequence if input has been a packed sequence.
         """
+
         if isinstance(x, rnn.PackedSequence) or lengths is None:
             assert lengths is None, "cannot combine x of type PackedSequence with lengths argument"
             return super().forward(x, hx=hx)
@@ -163,7 +200,7 @@ class RNN(ABC, nn.RNNBase):
                 return out, hidden_state
 
 
-class LSTM(RNN, nn.LSTM):
+class _LSTM(_RNN, nn.LSTM):
     """LSTM that can handle zero-length sequences"""
 
     def handle_no_encoding(self,
@@ -206,7 +243,9 @@ class LSTM(RNN, nn.LSTM):
         return hidden, cell
 
 
-class TimeDistributedEmbeddingBag(nn.EmbeddingBag):
+class _TimeDistributedEmbeddingBag(nn.EmbeddingBag):
+    """Used for categorical embedding"""
+
     def __init__(self,
                  *args,
                  batch_first: bool = False,
@@ -233,7 +272,9 @@ class TimeDistributedEmbeddingBag(nn.EmbeddingBag):
         return y
 
 
-class MultiEmbedding(nn.Module):
+class _MultiEmbedding(nn.Module):
+    """Can be used to do multi-embed categorical variables"""
+
     def __init__(self,
                  embedding_sizes: Dict[str, Tuple[int, int]],
                  categorical_groups: Dict[str, List[str]],
@@ -259,7 +300,7 @@ class MultiEmbedding(nn.Module):
             self.embedding_sizes[name][1] = embedding_size
 
             if name in self.categorical_groups:  # embedding bag if related embeddings
-                embeddings[name] = TimeDistributedEmbeddingBag(
+                embeddings[name] = _TimeDistributedEmbeddingBag(
                     self.embedding_sizes[name][0], embedding_size, mode="sum", batch_first=True
                 )
             else:
@@ -301,11 +342,9 @@ class MultiEmbedding(nn.Module):
         return input_vectors
 
 
-class TimeDistributedInterpolation(nn.Module):
-    """==========================================CHECKED==========================================
-    interpolates input size to output size.
-    This is lke TimeDistributed with interpolation
-    """
+class _TimeDistributedInterpolation(nn.Module):
+    """interpolates input size to output size."""
+
     def __init__(self,
                  output_size: int,
                  batch_first: bool = False):
@@ -336,17 +375,22 @@ class TimeDistributedInterpolation(nn.Module):
         return y
 
 
-class GatedLinearUnit(nn.Module):
+class _GatedLinearUnit(nn.Module):
     def __init__(self,
                  input_size: int,
                  hidden_size: int = None,
                  dropout: float = None):
 
         """Applies a Gated Linear Unit (GLU) to an input, see equation (5).
-            Args:
-                input_size: input size of x
-                hidden_size: Dimension of GLU
-                dropout: Dropout rate to apply if any
+
+        Parameters
+        ----------
+        input_size
+            input size of x
+        hidden_size
+            Dimension of GLU
+        dropout
+            Dropout rate to apply if any
         """
 
         super().__init__()
@@ -386,11 +430,8 @@ class GatedLinearUnit(nn.Module):
         return x
 
 
-class ResampleNorm(nn.Module):
-    """==========================================SEMI-CHECKED==========================================
-    Added option USE_ADAPTION_RESAMPLE to see if it's better just with TimeDistributedInterpolation.
-    Resamples an input to an output size
-    """
+class _ResampleNorm(nn.Module):
+    """Resamples an input to an output size"""
 
     def __init__(self,
                  input_size: int,
@@ -402,7 +443,7 @@ class ResampleNorm(nn.Module):
         self.output_size = output_size or input_size
 
         if self.input_size != self.output_size:
-            self.resample = TimeDistributedInterpolation(self.output_size, batch_first=True)
+            self.resample = _TimeDistributedInterpolation(self.output_size, batch_first=True)
 
         self.mask = nn.Parameter(torch.zeros(self.output_size, dtype=torch.float))
         self.gate = nn.Sigmoid()
@@ -418,12 +459,12 @@ class ResampleNorm(nn.Module):
         return output
 
 
-class AddNorm(nn.Module):
+class _AddNorm(nn.Module):
+    """Applies skip connection followed by layer normalisation."""
+
     def __init__(self,
                  input_size: int,
                  skip_size: int = None):
-        """Applies skip connection followed by layer normalisation.
-        """
 
         super().__init__()
 
@@ -431,7 +472,7 @@ class AddNorm(nn.Module):
         self.skip_size = skip_size or input_size
 
         if self.input_size != self.skip_size:
-            self.resample = TimeDistributedInterpolation(self.input_size, batch_first=True)
+            self.resample = _TimeDistributedInterpolation(self.input_size, batch_first=True)
 
         self.norm = nn.LayerNorm(self.input_size)
 
@@ -445,8 +486,9 @@ class AddNorm(nn.Module):
         return output
 
 
-class GateAddNorm(nn.Module):
+class _GateAddNorm(nn.Module):
     """Equation (2)"""
+
     def __init__(self,
                  input_size: int,
                  hidden_size: int = None,
@@ -460,8 +502,8 @@ class GateAddNorm(nn.Module):
         self.skip_size = skip_size or self.hidden_size
         self.dropout = dropout
 
-        self.glu = GatedLinearUnit(self.input_size, hidden_size=self.hidden_size, dropout=self.dropout)
-        self.add_norm = AddNorm(self.hidden_size, skip_size=self.skip_size)
+        self.glu = _GatedLinearUnit(self.input_size, hidden_size=self.hidden_size, dropout=self.dropout)
+        self.add_norm = _AddNorm(self.hidden_size, skip_size=self.skip_size)
 
     def forward(self, x, skip):
         # skip is the same as residual
@@ -470,7 +512,7 @@ class GateAddNorm(nn.Module):
         return output
 
 
-class GatedResidualNetwork(nn.Module):
+class _GatedResidualNetwork(nn.Module):
     """Top right graph in Figure 2 and formulas (2) -- (5)"""
 
     def __init__(self,
@@ -480,11 +522,17 @@ class GatedResidualNetwork(nn.Module):
                  dropout: float = 0.1,
                  context_size: Optional[int] = None):
         """Applies the gated residual network (GRN) as defined in paper.
-        Args:
-            hidden_size: Internal state size
-            output_size: Size of output layer
-            dropout: Dropout rate if dropout is applied
-            context_size: size of optional context vector
+
+        Parameters
+        ----------
+        hidden_size
+            Internal state size
+        output_size
+            Size of output layer
+        dropout
+            Dropout rate if dropout is applied
+        context_size
+            size of optional context vector
         """
 
         super().__init__()
@@ -497,9 +545,9 @@ class GatedResidualNetwork(nn.Module):
         # convert raw input into output size for residuals
         if self.input_size != self.output_size:
             if not USE_ADAPTION_RESAMPLE:
-                self.resample_norm = ResampleNorm(self.input_size, self.output_size)
+                self.resample_norm = _ResampleNorm(self.input_size, self.output_size)
             else:
-                self.resample_norm = TimeDistributedInterpolation(self.output_size, batch_first=True)
+                self.resample_norm = _TimeDistributedInterpolation(self.output_size, batch_first=True)
 
         self.fc1 = nn.Linear(self.input_size, self.hidden_size)
         if self.context_size is not None:
@@ -511,7 +559,7 @@ class GatedResidualNetwork(nn.Module):
 
         self.init_weights()
 
-        self.gate_add_norm = GateAddNorm(
+        self.gate_add_norm = _GateAddNorm(
             input_size=self.hidden_size,
             skip_size=self.output_size,
             hidden_size=self.output_size,
@@ -529,9 +577,7 @@ class GatedResidualNetwork(nn.Module):
                 torch.nn.init.xavier_uniform_(p)
 
     def forward(self, x, context=None):
-        """seems to be correct"""
-
-        # residual, also called `skip` basically bypasses until add_norm
+        # residual connection, also called `skip` basically bypasses until add_norm
         if self.input_size != self.output_size:
             residual = self.resample_norm(x)
         else:
@@ -546,18 +592,23 @@ class GatedResidualNetwork(nn.Module):
         return x
 
 
-class VariableSelectionNetwork(nn.Module):
+class _VariableSelectionNetwork(nn.Module):
     def __init__(self,
                  input_sizes: Dict[str, int],
                  hidden_size: int,
                  input_embedding_flags: Optional[Dict[str, bool]] = None,
                  dropout: float = 0.1,
                  context_size: Optional[int] = None,
-                 single_variable_grns: Optional[Dict[str, GatedResidualNetwork]] = None,
+                 single_variable_grns: Optional[Dict[str, _GatedResidualNetwork]] = None,
                  prescalers: Optional[torch.nn.ModuleDict] = None):
         """
-        Calculate weights for ``num_inputs`` variables  which are each of size ``input_size``
+        From paper https://arxiv.org/pdf/1912.09363.pdf:
+        TFT is designed to provide instance-wise variable selection through the use of variable selection networks
+        applied to both static covariates and time-dependent covariates. Beyond provid- ing insights into which
+        variables are most significant for the prediction problem, variable selection also allows TFT to remove any
+        unnecessary noisy inputs which could negatively impact performance.
         """
+
         super().__init__()
 
         self.input_sizes = input_sizes
@@ -573,7 +624,7 @@ class VariableSelectionNetwork(nn.Module):
 
         if self.num_inputs >= 1:
             # right side of figure 2 bottom right graph
-            self.vars_flattened_grn = GatedResidualNetwork(
+            self.vars_flattened_grn = _GatedResidualNetwork(
                 input_size=self.input_sizes_total,
                 hidden_size=min(self.hidden_size, self.num_inputs) if not USE_ADAPTION_VSN else self.hidden_size,
                 output_size=self.num_inputs,
@@ -581,6 +632,7 @@ class VariableSelectionNetwork(nn.Module):
                 context_size=self.context_size
             )
 
+        # TODO: (dbader) this can be cleaner, also think about moving prescalers out of _VariableSelectionNetwork
         # left side of figure 2 bottom right graph
         self.vars_single_grns = nn.ModuleDict()
         self.prescalers = nn.ModuleDict()
@@ -589,11 +641,11 @@ class VariableSelectionNetwork(nn.Module):
                 self.vars_single_grns[name] = single_variable_grns[name]
             elif self.input_embedding_flags.get(name, False):
                 if not USE_ADAPTION_RESAMPLE:
-                    self.vars_single_grns[name] = ResampleNorm(input_size, self.hidden_size)
+                    self.vars_single_grns[name] = _ResampleNorm(input_size, self.hidden_size)
                 else:
-                    self.vars_single_grns[name] = TimeDistributedInterpolation(self.hidden_size, batch_first=True)
+                    self.vars_single_grns[name] = _TimeDistributedInterpolation(self.hidden_size, batch_first=True)
             else:
-                self.vars_single_grns[name] = GatedResidualNetwork(
+                self.vars_single_grns[name] = _GatedResidualNetwork(
                     input_size=input_size,
                     hidden_size=min(input_size, self.hidden_size) if not USE_ADAPTION_VSN else self.hidden_size,
                     output_size=self.hidden_size,
@@ -633,8 +685,8 @@ class VariableSelectionNetwork(nn.Module):
         return outputs, selection_weights
 
 
-class ScaledDotProductAttention(nn.Module):
-    """ScaledDotProductAttention is an self-attention mechanism to learn long-term
+class _ScaledDotProductAttention(nn.Module):
+    """_ScaledDotProductAttention is a self-attention mechanism to learn long-term
     relationships across different time steps. It scales values V based on relationship
     between Keys K and queries Q.
 
@@ -642,17 +694,22 @@ class ScaledDotProductAttention(nn.Module):
 
     def __init__(self):
 
-        super(ScaledDotProductAttention, self).__init__()
+        super(_ScaledDotProductAttention, self).__init__()
 
         self.softmax = nn.Softmax(dim=2)
 
     def forward(self, q, k, v, mask=None):
         """
-        Attributes:
-            q: Queries matrix of dimension (N x d_attention)
-            k: Keys matrix of dimension (N x d_attention)
-            v: Values matrix of dimension (N x d_Values)
-            mask: masking if required -- sets softmax to very large value
+        Parameters
+        ----------
+        q
+            Queries matrix of dimension (N x d_attention)
+        k
+            Keys matrix of dimension (N x d_attention)
+        v
+            Values matrix of dimension (N x d_Values)
+        mask
+            masking if required -- sets softmax to very large value
         """
 
         attn = torch.bmm(q, k.transpose(1, 2))
@@ -668,7 +725,7 @@ class ScaledDotProductAttention(nn.Module):
         return output, attn
 
 
-class InterpretableMultiHeadAttention(nn.Module):
+class _InterpretableMultiHeadAttention(nn.Module):
     """Equations (13) -- (16)"""
 
     def __init__(self,
@@ -676,7 +733,7 @@ class InterpretableMultiHeadAttention(nn.Module):
                  d_model: int,
                  dropout: Optional[float] = None):
         """
-        Attributes:
+        Description of parameters
             n_head: number of heads
             d_model: TFT state dimensionality
             d_k: Key/query dimensionality per head
@@ -688,7 +745,7 @@ class InterpretableMultiHeadAttention(nn.Module):
             w_h: output head weight matrix to project internal state to the original TFT state size
         """
 
-        super(InterpretableMultiHeadAttention, self).__init__()
+        super(_InterpretableMultiHeadAttention, self).__init__()
 
         self.n_head = n_head
         self.d_model = d_model
@@ -699,8 +756,8 @@ class InterpretableMultiHeadAttention(nn.Module):
         self.v_layer = nn.Linear(self.d_model, self.d_v)
         self.q_layers = nn.ModuleList([nn.Linear(self.d_model, self.d_q) for _ in range(self.n_head)])
         self.k_layers = nn.ModuleList([nn.Linear(self.d_model, self.d_k) for _ in range(self.n_head)])
-        # use same value layer to facilitate interpretation
-        self.attention = ScaledDotProductAttention()
+
+        self.attention = _ScaledDotProductAttention()
         self.w_h = nn.Linear(self.d_v, self.d_model, bias=False)
 
         self.init_weights()
@@ -720,9 +777,8 @@ class InterpretableMultiHeadAttention(nn.Module):
             qs = self.q_layers[i](q)
             ks = self.k_layers[i](k)
             head, attn = self.attention(qs, ks, vs, mask)
-            if not USE_ADAPTION_NO_ATTENTION_DROPOUT:
-                if self.dropout is not None:
-                    head = self.dropout(head)
+            if self.dropout is not None:
+                head = self.dropout(head)
             heads.append(head)
             attns.append(attn)
 
@@ -732,8 +788,7 @@ class InterpretableMultiHeadAttention(nn.Module):
         outputs = torch.mean(head, dim=2) if self.n_head > 1 else head
 
         outputs = self.w_h(outputs)
-        if not USE_ADAPTION_NO_ATTENTION_DROPOUT:
-            if self.dropout is not None:
-                outputs = self.dropout(outputs)
+        if self.dropout is not None:
+            outputs = self.dropout(outputs)
 
         return outputs, attn
